Bug 1735899 - Make sure RemoteLazyInputStream::Close calls mInputStreamCallback OnInputStreamReady method. r=nika
RemoteLazyInputStream::Close was not setting mInputStreamCallback to a nullptr without using it in the
InputStreamCallbackRunnable::Execute, which would be calling OnInputStreamReady on the mInputStreamCallback.

This does also match the details we gathered while investigating the bug (which was triggered exactly by a remote
lazy stream getter that did never got to call OnInputStreamReady by the time we were closing the channel).

Differential Revision: https://phabricator.services.mozilla.com/D129374